third_party/blink/web_tests/platform/generic/fast/dom/HTMLImageElement/image-sizes-js-change-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS currentSrcFileName("foo") is "image-set-4x.png"
PASS internals.isLoadingFromMemoryCache("resources/image-set-4x.png") is true
PASS internals.isLoadingFromMemoryCache("../../hidpi/resources/image-set-2x.png") is false
PASS internals.isLoadingFromMemoryCache("../../hidpi/resources/image-set-1x.png") is false
PASS internals.isLoadingFromMemoryCache("resources/blue-100-px-square.png") is false
